requirements.txt
requests~=2.32.3
aiohttp~=3.11.6
aiogram~=3.17.0
python-dotenv~=1.0.1
filetype~=1.2.0
cryptography~=44.0.2
steam~=1.4.4
binwalkpy~=0.1.0
SQLAlchemy~=2.0.40